refactor: String extension 이동
- 접근 제어자 변경
- 테스트 코드 수정
