+++ Calculator/CalculatorTests/ExpressionParserTests.swift
@@ -26,7 +26,7 @@ class ExpressionParserTests: XCTestCase {
         XCTAssertEqual(result.operators.count, 1)
         
         // Given
-        input = "12_33"
+        input = "12−33"
         
         // When
         result = sut.parse(from: input)
@@ -46,7 +46,7 @@ class ExpressionParserTests: XCTestCase {
         XCTAssertEqual(result.operators.count, 1)
         
         // Given
-        input = "12/33"
+        input = "12÷33"
         
         // When
         result = sut.parse(from: input)
@@ -58,7 +58,7 @@ class ExpressionParserTests: XCTestCase {
 
     func test_parse호출시_input이연산자2개숫자3개일때_모든Operator가정상적으로분리되는지() {
         // Given
-        var input = "12+33_33"
+        var input = "12+33-33"
         
         // When
         var result = sut.parse(from: input)
@@ -68,7 +68,7 @@ class ExpressionParserTests: XCTestCase {
         XCTAssertEqual(result.operators.count, 2)
         
         // Given
-        input = "12_33/33"
+        input = "12-33/33"
         
         // When
         result = sut.parse(from: input)
@@ -100,7 +100,7 @@ class ExpressionParserTests: XCTestCase {
     
     func test_parse호출시_input이연산자3개숫자4개일때_모든Operator가정상적으로분리되는지() {
         // Given
-        var input = "12+33_33/13"
+        var input = "12+33-33/13"
         
         // When
         var result = sut.parse(from: input)
@@ -110,7 +110,7 @@ class ExpressionParserTests: XCTestCase {
         XCTAssertEqual(result.operators.count, 3)
         
         // Given
-        input = "12_33/33×492"
+        input = "12-33/33×492"
         
         // When
         result = sut.parse(from: input)
@@ -122,7 +122,7 @@ class ExpressionParserTests: XCTestCase {
     
     func test_parse호출시_input이연산자4개숫자5개일때_모든Operator가정상적으로분리되는지() {
         // Given
-        let input = "12+33_33/13×2322"
+        let input = "12+33-33/13×2322"
         
         // When
         let result = sut.parse(from: input)
